+++ src/modules/Privacy/index.tsx
@@ -12,6 +12,7 @@ import {
   IconCookie,
   IconLock,
   IconPhone,
+  IconPoint,
   IconSpy,
   IconUsers,
 } from "@tabler/icons-react";
@@ -26,8 +27,6 @@ const PagePrivacy = () => {
         <Title ta="center" mb="xl">
           Privacy Policy
         </Title>
-
-        <Text>Last updated: April 24, 2024</Text>
 
         <Flex mt="xl" mb="sm" align="center">
           <Text fz={26} fw={600}>
@@ -60,123 +59,96 @@ const PagePrivacy = () => {
         </Text>
 
         <Flex mt="xl" mb="sm" align="center">
-          <IconUsers style={getIconStyle(32)} color={theme.colors.pink["5"]} />
+          <IconUsers style={getIconStyle(32)} color={theme.colors.grape["5"]} />
 
           <Text fz={26} fw={600} ml="xs">
             3rd parties
           </Text>
         </Flex>
 
-        <List>
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              Cloudflare:
-            </Text>{" "}
-            <Text component="span">
-              Right now Khofly is hosted on Cloudflare Pages, this means that
-              Cloudflare has access to all incoming requests, their privacy
-              policy can be found{" "}
-              <Anchor
-                href="https://www.cloudflare.com/privacypolicy/"
-                target="_blank"
-                rel="noreferrer noopener"
-              >
-                <Text component="span" c="blue.4">
-                  here
-                </Text>
-              </Anchor>
+        <Text>
+          <strong>Cloudflare:</strong> Right now Khofly is hosted on Cloudflare
+          Pages, this means that Cloudflare has access to all incoming requests,
+          their privacy policy can be found{" "}
+          <Anchor
+            href="https://www.cloudflare.com/privacypolicy/"
+            target="_blank"
+            rel="noreferrer noopener"
+          >
+            <Text component="span" c="blue.4">
+              here
             </Text>
-          </List.Item>
+          </Anchor>
+        </Text>
 
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              OpenStreetMaps:
-            </Text>{" "}
-            <Text component="span">
-              Public Nominatim API by OpenStreetMaps is used for Geocoding data
-              in maps, their privacy policy can be found{" "}
-              <Anchor
-                href="https://osmfoundation.org/wiki/Privacy_Policy"
-                target="_blank"
-                rel="noreferrer noopener"
-              >
-                <Text component="span" c="blue.4">
-                  here
-                </Text>
-              </Anchor>
+        <Text component="span">
+          <strong>OpenStreetMaps:</strong> Public Nominatim API by
+          OpenStreetMaps is used for Geocoding data in maps, their privacy
+          policy can be found{" "}
+          <Anchor
+            href="https://osmfoundation.org/wiki/Privacy_Policy"
+            target="_blank"
+            rel="noreferrer noopener"
+          >
+            <Text component="span" c="blue.4">
+              here
             </Text>
-          </List.Item>
+          </Anchor>
+        </Text>
 
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              Hetzner:
-            </Text>{" "}
-            <Text component="span">
-              SearXNG instance that Khofly uses is hosted on a Hetzner VPS,
-              their privacy policy can be found{" "}
-              <Anchor
-                href="https://www.hetzner.com/legal/privacy-policy/"
-                target="_blank"
-                rel="noreferrer noopener"
-              >
-                <Text component="span" c="blue.4">
-                  here
-                </Text>
-              </Anchor>
+        <Text component="span">
+          <strong>Hetzner:</strong> SearXNG instance that Khofly uses is hosted
+          on a Hetzner VPS, their privacy policy can be found{" "}
+          <Anchor
+            href="https://www.hetzner.com/legal/privacy-policy/"
+            target="_blank"
+            rel="noreferrer noopener"
+          >
+            <Text component="span" c="blue.4">
+              here
             </Text>
-          </List.Item>
+          </Anchor>
+        </Text>
 
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              DuckDuckGo:
-            </Text>{" "}
-            <Text component="span">
-              If you decide to display website favicons in settings this will
-              send a request to DDGs servers, keep in mind that this option is
-              disabled by default, their privacy policy can be found{" "}
-              <Anchor
-                href="https://duckduckgo.com/privacy"
-                target="_blank"
-                rel="noreferrer noopener"
-              >
-                <Text component="span" c="blue.4">
-                  here
-                </Text>
-              </Anchor>
+        <Text component="span">
+          <strong>DuckDuckGo:</strong> If you decide to display website favicons
+          in settings this will send a request to DDGs servers, keep in mind
+          that this option is disabled by default, their privacy policy can be
+          found{" "}
+          <Anchor
+            href="https://duckduckgo.com/privacy"
+            target="_blank"
+            rel="noreferrer noopener"
+          >
+            <Text component="span" c="blue.4">
+              here
             </Text>
-          </List.Item>
-        </List>
+          </Anchor>
+        </Text>
 
         <Flex mt="xl" mb="sm" align="center">
-          <IconLock style={getIconStyle(32)} color={theme.colors.green["6"]} />
+          <IconLock style={getIconStyle(32)} color={theme.colors.green["5"]} />
 
           <Text fz={26} fw={600} ml="xs">
             For improved privacy
           </Text>
         </Flex>
 
-        <List>
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              self-host Khofly
-            </Text>
-          </List.Item>
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              change/self-host SearXNG instance
-            </Text>
-          </List.Item>
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              change/self-host Nominatim instance
-            </Text>
-          </List.Item>
-          <List.Item>
-            <Text component="span" fz={18} fw={700}>
-              disable website favicons in settings
-            </Text>
-          </List.Item>
-        </List>
+        <Text component="span" fw={700}>
+          self-host Khofly
+        </Text>
+
+        <Text component="span" fw={700}>
+          change/self-host SearXNG instance
+        </Text>
+
+        <Text component="span" fw={700}>
+          change/self-host Nominatim instance
+        </Text>
+
+        <Text component="span" fw={700}>
+          disable website favicons in settings
+        </Text>
 
         <Text mt="md">
           While you can generally trust public instances the only way to truly
@@ -201,7 +173,7 @@ const PagePrivacy = () => {
         </Text>
 
         <Text>
-          <b>Email:</b>{" "}
+          <strong>Email:</strong>{" "}
           <Anchor
             href="https://discord.gg/mQ68HppVbt"
             target="_blank"
@@ -213,7 +185,7 @@ const PagePrivacy = () => {
           </Anchor>
         </Text>
         <Text>
-          <b>Discord:</b>{" "}
+          <strong>Discord:</strong>{" "}
           <Anchor
             href="https://discord.gg/mQ68HppVbt"
             target="_blank"
@@ -224,6 +196,8 @@ const PagePrivacy = () => {
             </Text>
           </Anchor>
         </Text>
+
+        <Text ta="right">Last updated: April 24, 2024</Text>
       </Card>
     </Container>
   );