+++ src/modules/Settings/components/General/ShowEnginesSwitch/index.tsx
@@ -1,0 +1,18 @@
+import { Switch } from "@mantine/core";
+import { useSettingsStore } from "@store/settings";
+
+const ShowEnginesSwitch = () => {
+  const { showEngines, setShowEngines } = useSettingsStore((state) => ({
+    showEngines: state.showEngines,
+    setShowEngines: state.setShowEngines,
+  }));
+
+  return (
+    <Switch
+      checked={showEngines}
+      onChange={(e) => setShowEngines(e.currentTarget.checked)}
+    />
+  );
+};
+
+export default ShowEnginesSwitch;